+++ src/clustermergeronesampletester.hpp
@@ -20,8 +20,8 @@ class ClusterMergerOneSampleTester : public ClusterMergerTester {
     
 public:
     ClusterMergerOneSampleTester(double zvalue, double error);
-    bool shouldMerge(const std::vector<std::array<int,4>>& freq_c1,
-                     const std::vector<std::array<int,4>>& freq_c2);
+    bool shouldMerge(const std::vector<std::array<uint64_t,4>>& freq_c1,
+                     const std::vector<std::array<uint64_t,4>>& freq_c2);
 private:
     std::unique_ptr<HypothesisTester> _tester;
 };
