+++ src/mixturebptester.cpp
@@ -11,7 +11,7 @@ MixtureBPTester::MixtureBPTester(double p_value, double error_rate)
   * TODO(lu): Finish this function. Try to use a hypothysis test instead of
   * the gating strategy.
 */
-bool MixtureBPTester::IsPotentialMixture(const std::array<int, 4>& frequency, size_t bp) {
+bool MixtureBPTester::IsPotentialMixture(const std::array<uint64_t, 4>& frequency, size_t bp) {
     int total = std::accumulate(frequency.begin(), frequency.end(), 0);
     return frequency[bp] * 1.0 / total > _error_rate;
     //return true;
