+++ src/split_util.h
@@ -8,7 +8,7 @@
 
 namespace barcodeSpace {
 
-inline double Entropy(const std::array<int, 4>& base_freq) {
+inline double Entropy(const std::array<uint64_t, 4>& base_freq) {
     double majority = *std::max_element(base_freq.begin(), base_freq.end());
 
     double sum = std::accumulate(base_freq.begin(), base_freq.end(), 0);
@@ -20,14 +20,14 @@ inline double Entropy(const std::array<int, 4>& base_freq) {
     }
 }
 
-inline std::vector<double> Entropy(const std::vector<std::array<int, 4>>& bpfrequencies) {
+inline std::vector<double> Entropy(const std::vector<std::array<uint64_t, 4>>& bpfrequencies) {
     std::vector<double> entropies;
     for (const auto bp_freq : bpfrequencies) {
         entropies.push_back(Entropy(bp_freq));
     }
     return entropies;
 }
-inline double Entropy(const std::array<int, 4>& base_freq, double* majority, double* total) {
+inline double Entropy(const std::array<uint64_t, 4>& base_freq, double* majority, double* total) {
     *majority = *std::max_element(base_freq.begin(), base_freq.end());
 
     *total = std::accumulate(base_freq.begin(), base_freq.end(), 0);
@@ -35,24 +35,25 @@ inline double Entropy(const std::array<int, 4>& base_freq, double* majority, dou
         return 0.0;
     } else {
         double ratio = *majority / *total;
-        return -1 * (ratio * log2(ratio) + (1-ratio)*log2(1-ratio));
+        return -1 * (ratio * log2(ratio) + (1 - ratio) * log2(1 - ratio));
     }
 }
 
-inline int majorityFrequency(const std::array<int, 4>& base_freq) {
+inline uint64_t majorityFrequency(const std::array<uint64_t, 4>& base_freq) {
     return *std::max_element(base_freq.begin(), base_freq.end());
     
 }
-inline long int majorityBP(const std::array<int, 4>& base_freq) {
+inline uint64_t majorityBP(const std::array<uint64_t, 4>& base_freq) {
     return std::max_element(base_freq.begin(), base_freq.end()) - base_freq.begin();
 }
 
-inline std::vector<int> center(const std::vector<std::array<int, 4>>& bpfrequencies) {
-    std::vector<int> center;
+inline std::vector<uint64_t> center(const std::vector<std::array<uint64_t, 4>>& bpfrequencies) {
+    std::vector<uint64_t> center;
     for(const auto& b : bpfrequencies) {
         center.push_back(majorityBP(b));
     }
     return center;
 }
+    
 }   // namespace barcodeSpace
 #endif // SPLIT_UTIL_H
